feat(SimpleList): move tests, add alternate callback props
